object-localization/code/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.7
google-cloud-bigquery>=3.9.0
google-cloud-vision>=3.4.1
